+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to CloudQuery will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [Unreleased] 
+
+### :rocket: Added
+* Added core migrations implementation
+* Added fetch summary saving to `fetches` table
 
 
 <!--